+++ shared/src/rusterix_utils.rs
@@ -5,6 +5,7 @@ use rusterix::{Command, Entity, Rusterix, Value};
 pub fn start_server(rusterix: &mut Rusterix, project: &mut Project) {
     rusterix.server.clear();
     rusterix.server.log_changed = true;
+    rusterix.assets.global = project.render_graph.clone();
 
     insert_content_into_maps(project);